+++ zingocli/tests/darkside/mod.rs
@@ -1,2 +1,3 @@
+mod constants;
 #[cfg(feature = "darksidewalletd_tests")]
 pub mod darksidewalletd;
