--- zingocli/tests/darksidewalletd/darksidewalletd.rs
@@ -1,283 +1,0 @@
-tonic::include_proto!("cash.z.wallet.sdk.rpc");
-
-use crate::utils::scenarios::setup::ClientManager;
-use darkside_streamer_client::DarksideStreamerClient;
-use json::JsonValue;
-use rand::Rng;
-use tempdir::{self, TempDir};
-use tokio::time::sleep;
-
-use std::sync::Arc;
-
-use http_body::combinators::UnsyncBoxBody;
-use hyper::{client::HttpConnector, Uri};
-use tokio_rustls::rustls::{ClientConfig, RootCertStore};
-use tonic::Status;
-use tower::{util::BoxCloneService, ServiceExt};
-
-type UnderlyingService = BoxCloneService<
-    http::Request<UnsyncBoxBody<prost::bytes::Bytes, Status>>,
-    http::Response<hyper::Body>,
-    hyper::Error,
->;
-
-macro_rules! define_darkside_connector_methods(
-    ($($name:ident (&$self:ident, $($param:ident: $param_type:ty),*$(,)?) {$param_packing:expr}),*) => {$(
-        #[allow(unused)]
-        pub(crate) async fn $name(&$self, $($param: $param_type),*) -> ::std::result::Result<(), String> {
-            let request = ::tonic::Request::new($param_packing);
-
-            let mut client = $self.get_client().await.map_err(|e| format!("{e}"))?;
-
-        let _response = client
-            .$name(request)
-            .await
-            .map_err(|e| format!("{}", e))?
-            .into_inner();
-
-        Ok(())
-    })*}
-);
-
-#[derive(Clone)]
-pub struct DarksideConnector(http::Uri);
-
-impl DarksideConnector {
-    pub fn new(uri: http::Uri) -> Self {
-        Self(uri)
-    }
-
-    pub(crate) fn get_client(
-        &self,
-    ) -> impl std::future::Future<
-        Output = Result<DarksideStreamerClient<UnderlyingService>, Box<dyn std::error::Error>>,
-    > {
-        let uri = Arc::new(self.0.clone());
-        async move {
-            let mut http_connector = HttpConnector::new();
-            http_connector.enforce_http(false);
-            if uri.scheme_str() == Some("https") {
-                let mut roots = RootCertStore::empty();
-                roots.add_server_trust_anchors(webpki_roots::TLS_SERVER_ROOTS.0.iter().map(
-                    |anchor_ref| {
-                        tokio_rustls::rustls::OwnedTrustAnchor::from_subject_spki_name_constraints(
-                            anchor_ref.subject,
-                            anchor_ref.spki,
-                            anchor_ref.name_constraints,
-                        )
-                    },
-                ));
-
-                let tls = ClientConfig::builder()
-                    .with_safe_defaults()
-                    .with_root_certificates(roots)
-                    .with_no_client_auth();
-                let connector = tower::ServiceBuilder::new()
-                    .layer_fn(move |s| {
-                        let tls = tls.clone();
-
-                        hyper_rustls::HttpsConnectorBuilder::new()
-                            .with_tls_config(tls)
-                            .https_or_http()
-                            .enable_http2()
-                            .wrap_connector(s)
-                    })
-                    .service(http_connector);
-                let client = Box::new(hyper::Client::builder().build(connector));
-                let uri = uri.clone();
-                let svc = tower::ServiceBuilder::new()
-                    //Here, we take all the pieces of our uri, and add in the path from the Requests's uri
-                    .map_request(move |mut req: http::Request<tonic::body::BoxBody>| {
-                        let uri = Uri::builder()
-                            .scheme(uri.scheme().unwrap().clone())
-                            .authority(uri.authority().unwrap().clone())
-                            //here. The Request's uri contains the path to the GRPC sever and
-                            //the method being called
-                            .path_and_query(req.uri().path_and_query().unwrap().clone())
-                            .build()
-                            .unwrap();
-
-                        *req.uri_mut() = uri;
-                        req
-                    })
-                    .service(client);
-
-                Ok(DarksideStreamerClient::new(svc.boxed_clone()))
-            } else {
-                let connector = tower::ServiceBuilder::new().service(http_connector);
-                let client = Box::new(hyper::Client::builder().http2_only(true).build(connector));
-                let uri = uri.clone();
-                let svc = tower::ServiceBuilder::new()
-                    //Here, we take all the pieces of our uri, and add in the path from the Requests's uri
-                    .map_request(move |mut req: http::Request<tonic::body::BoxBody>| {
-                        let uri = Uri::builder()
-                            .scheme(uri.scheme().unwrap().clone())
-                            .authority(uri.authority().unwrap().clone())
-                            //here. The Request's uri contains the path to the GRPC sever and
-                            //the method being called
-                            .path_and_query(req.uri().path_and_query().unwrap().clone())
-                            .build()
-                            .unwrap();
-
-                        *req.uri_mut() = uri;
-                        req
-                    })
-                    .service(client);
-
-                Ok(DarksideStreamerClient::new(svc.boxed_clone()))
-            }
-        }
-    }
-
-    define_darkside_connector_methods!(
-        apply_staged(&self, height: i32) { DarksideHeight { height } },
-        add_tree_state(&self, tree_state: TreeState) { tree_state },
-        reset(
-            &self,
-            sapling_activation: i32,
-            branch_id: String,
-            chain_name: String,
-        ) {
-            DarksideMetaState {
-                sapling_activation,
-                branch_id,
-                chain_name,
-            }
-        },
-        stage_blocks(&self, url: String) { DarksideBlocksUrl { url } },
-        stage_blocks_stream(&self, blocks: Vec<String>) { ::futures_util::stream::iter(blocks.into_iter().map(|block| DarksideBlock { block })) }
-    );
-}
-
-async fn prepare_darksidewalletd(uri: http::Uri) -> Result<(), String> {
-    let connector = DarksideConnector(uri);
-
-    // reset with parameters
-    connector
-        .reset(663_150, String::from("2bb40e60"), String::from("main"))
-        .await?;
-
-    connector.stage_blocks_stream(vec![String::from("0400000066f5b2da725bc918701fcf0ded2ffb5cd4a1ad767e61a436d6e0b543a1084f0cabeae3d896f391490a90c5a3252ab0a69b838d9cf44247ab468ad295abe110516e01d4126cd47c1e84066d5213783b097da9a69a76fa0af67a3d30fd13272b1cfe487e640f0f0f201700280b680f52ce6df51b8112e3262630cbab478f5facbbf064490c43e60000241e70b21ff1fcde47f528da146f37e47ab59d2322631c839c85935635dd901c03ad8a07a2020400008085202f89010000000000000000000000000000000000000000000000000000000000000000ffffffff03530101ffffffff000000000003000000b01abfdaffffffff0001304bf99fb152350553606da830567886e313eac3647f9a43375bc77c26baa53768f4daa341588705c0b57cc8052b5a9ccc78cf991bc5c18dbda992c8e19ec006917e71e775fcb4e12b4c27000cbf118090219db939b21400f4be76d35dd039c0d3846495b7e98a34cf63c224968f479e8bb1dd1af06cbb1d7a15db2aaf82ac54b5bc11ae0381b2df4dd21dc3a7d1e193bbdb8bcaf3beebba063148ecfd37e2cfecb57f0b9d2006e9d5f43bf0618f188d8f9583a279fbe0acc004ebd7f6e7b9ec1b999e9058a70546cc2a84a516e8a9bfe9a2bfb00bd9ce84fc2416859d18c688c4ade3405111e6e92c78fbd4828970eff9be5ecca16dca4424a805e0664b6fcf20ec96180b9e88f65b80bebec51d6c4c6bf3ddc8dfacd7ba55ecf06d20cc5cdc374e2605908fac1eccb1a4d93959d6cf60a0b2f6d54ecce17478db309b924d399b07f063c89a6fc9474e5c01e92d2adca7479cdf737489f860fcf847edf8b9041f8837c4cd15b31aabe2adcfe777cd224388c90dfccf7bc7a3af8d6a028737c3668014a422992270421660256c4dc2a7a6515c95bd50410a4cad31802b327f08a442f862c5de981e462f36bf581daf80b4c4ad720046991d650687da3704c92932a095118370ecc2c360557cdc09c201329fdeba5ef524479351a451c900ae23e2792b310e30217ae6a2954a346e162106a2555bde2c55d5df4387097a15c7be00381719ed9791b00b3d374ca454d92581c4a07f4e4ec0a60de50fd5dc6d120fdc3adcbdbf198cf6e28cc5dc39718398aa5c4b761ed8f16d1fa370648acd371b1ab0ad29c8eafe9e1c7dcacf603a81a0234e22f752b142cea8cc9fab6edcf49327fb5e6631ee2976f333118ed87d338f3c532d0610cdd76c5c806e2b8b4f178e57c987e2cbc24ea48fc51922b239754da7e3752befc25479d8bd7788e32f2b581d5c7bc54b6496a4651df2e279c46a3f3112e05d2a5df0d789b7ce151f0fbf641b38589d53913e5b95f00d02f5a8ea5c9a9bd0b519f75947dccb5835a66d0dad85aee8d967eb56994c0df03b95b6f2da033270089a00109d2014244688fdfcc0906e10c74aa02de9ab02dc17a0cf40e5c7051eb35c20d770c48af36b3027a151dd3ba134ace4dc3511ef1ac07f1b0512b7687d9beaa79412d12ee752f7dd17f226a2d2c38611fa32ef15bd420bc1451a1e4191ab1502d35e5a8d8e4e39219ac24e987b032afcfa01519a3622105854c767d8042ac3316490adbad5b814a64ad3b86645208443edae59f81dc1ae11229cc41923855e0e1ea58e03ea20b50be79f0d344c474f08808d2eb2850834695bf859596085004a6378b404720d02c79536d0bb96f77175cc05c87ae010a35665e82e1217e0a0ed4014ca3c15762aa3c843139992b04056357661cfa1b5a2e1ef5369ebb3bf00050000800a27a726b4d0d6c200000000170000000000010439005c48bba1a9cab6149b8ca6a1f6132e3aa8d28e1d7eddf5abbcc6c79a1358d9279fbb55289ac77f172f589d97d6b2dac8029ca535a096692771fc69f9007eb43b6a3d3c140b494fc886494edb461a184962c915212f2ad746bdc8a8dd2602ca42c22b1ac5aab7e14b7b8339debdde43132dfc6049859e192ed87344b9b4c398e91959fe1ff3466daebed4f0b4e1ddcdd60178155f9a034c6fc40c96959159a0c8eddcc89c82a41cb9dabf952950a020a5645b682b4271d8a2fea77cd18f6f0b977d058ba461d3d9b1859e43595c752a6822468287526e2a8e86d3378bcc216cb0cf4439bb20a13b94e6de1c8713436d2d7a776d378de272edc7e5377344726708e3da541afe828a911aef9fc80e7f129a8ac74d86e0e93e42c6c78d0ee7a8d62396a2883ca949e9f9ed9a6833aca3102706fd9d8d56dd154cc0f30bad4c5ee07bd6762ec548abd0802ce8cef20412e3fff7e6bf5ae731e27470e2d51929e03985bf59513faa5b2569af30feaf88593b7f98b512d196e924acb2e2c0146f6078c249f0e15e79ea8e8559d2525854fe89bf807ae5a82e6515742f8cac18dd74e7a58030ac105418d0a9ca8ff35b71373085e5767b208ec976e7f7f0ae69af22ebc4844f33d30fb03326db1fad08140fc80441f31a0118b593f5acf55bfa881c5ab4d4f12ff13613ea98912936ee3119cd0e444731e51fdce1259f318eb4ce05e14496c727032e6dc88dbdcb0d4890fc1b62fb81c0bcb931a320d15be674627847c8cc0df150c428215fc27706e88826ab8599208542e856af1408c3ab31fe7058b4f08591140cbbb1432c2abe87f114b09638d443259348128e7f04d82f756a815b606d260b8883673b533a71f25dc90a688a30b1bf926a5d0bea47ccedb04b5a0f7baecda9abecbc42a95f80c667ede22e2b488a8f496e7026d54d8da87f7c3cfd30b7c01ef95560fd4436774e3c62db31225af4b1f7685df3020441ed927ee0d4f6d31bffc97f7a1122ecf4b7dced1eeaf86673a4e023b587c19bb7df56e97ea3b6f37b17391ac460e730304029a33b667fe122622567a7494171050a2917926edfd6cd3ec670c343eabc1d956b8c142b6168a0ed00845336693d3252d992b692c304d784b40f8d68901daccfbf46dfe3f9f888cd007c1260e5daf9825d1178528638d73dbc6f9cca8c095534b259584d350fef1aebdb85e91784a7908f4406d42cfd18f068129bd109842eeb437a239768ec3a23c8624851668800c12774e3b90572580ec7a8959d8a0036463a3838249d291875542f18747b2fbef47a18ed86d131dd03d116cc31e82bb097539a0fee2115deeb6668978a5656ee8fa20308a8e4238e21d158fb66ab8582c34fa955d2cd949a56af7b29dd10e2c838becd5ed7b36db1225853574e45d7da02064cd2d2ed07c6c9c252249fb6fbaff3b78a646bd2d0036b74cc688ed88b21d8434c6b6897254230606d3d1179ba00ac20149122741f1a21a1d8ea7ee01abfd57db40f07fdc7b9333bcec3e2ee0e3b4af695ef2d679c7e82bc365c3b57bbb58572e793f298b18fe443c54fa4785433176f31ecf14733c087200d76254f08f2af87e364cd6f79d8a1f7d57b3056848f82aab163b88cf64293a81a837bd9f153536b1b5b69d87c606101cdd2ac68bd9baba017b1b0aa03c0fda57a0f3bbc8e8e2e2d9ee74ec332388236a9a2f555dde99a331b643594409380853cdabfa6985cc2ceec36f5e2e1fafe4bde4a42cb352b4c389fc8fd00456b43b5a532a9ce1b0aa129eaa27abc4785a2ded41065e41b28f45a60233d9204af461c01ab2b53d2192aac8857366ef9a06fd346e8bb092f549b24f72a148c5accbf2bb15f287866315afddb34b3afbc9faa07acbafe977600493412441ec599ed6da5a7c46ebf0cc3ae0ec51b12ddb655313036b6c2c745c92f2a1f1e8e4621cdd85d74aa9f81fb7af53761b96c1ebd4418fae006e2a7a46b1d1bfe77eb6a85a955f904b41c2d967e2e9a239b7832c475ad622083de954b78eca7013fedfaac7093fcdb139f89fccdf986462f1a77e8e2e79405bae124a8ff043a8bf8fb6683344ee746cc001c6aeb281aaa4cf30b83a6bfe3bb03c2c910f1a73cc9b5cca38f418507411752a248dae796083cb4434963c77bed1b830f14a401d2820e3d4b70d1e6429441059c7ddc0dc42240b942f9d598ed2f33def1410aced8c0bab1bd7a0764e7bfb563ab4f529ba1791ccdd3336a7f77114e25d8b9e55d195e837786beba45fbf540be402500000000c808ec0485036616d4853310c57a304c32695e423680687fc32a15dcc9485122a5c87b5b90c99c754f42896562a20d4e21cfa39fb5e87c05521d62cddb2d630626e3d3ec006bbd0cff7a4ed676cbc5bfadeee81140d6849790eb60117806109c3e8c505323d8a37d3b29a7080cd99373b11400907a3266fb278f92c8fe758db2081e5e946bc08ca9333af0b6e7cc25dd6bf90ac018181333d77f329b9cf3ba038ba6e0c4dfbc09baf9752100eabcd334ae70723cf59317c82ef88b6da298fa6aae2a7a0b94a48533333e21c59152b0e44a408aea9a0e124bfdf882f8101b2c08c2453c9c88ac36705bde33ba4c211e5f3649c9aac0f5f65f8432904c164cd502f6f042c19b9e89f6f96adcc03974594ddc4c0d93f3d8e6aebee371f5e77d8b0db19874ebc981a519c044e0cfbe193205c055945da74a3deeb41255e1ea1d05215cb6a4adf54efc8ea610ff61de6d04a68f0429c0dd7af468c13a97c0f6325ead0c2d3cfe2c95265cd131ae4f7944771e3fd1e9e1c58f571f1235561a1181b70618299535c4f2dc1e89a578eaa3fb964b8e61dea25034105f77c1fb896d0e5937ad2313970651b4b7491f9efc436c7cd098bf78e92676435c280cccd7943c8dc6a5848e5f29f530bfa8e1e35b5e71482a7ef71ca732b40f08a5c06f795df92583829f817ff5b937c1632c0ff79dbb0cdc42f54bd787625334382677f1ee75e52d1c14a9af9916180c712fa4466953dad3981fa9d56df7e915b4ceb34047d022ad92496c06100215a26ccd43d55801c90121620b2a95b775a4412cae6f29cb11db0afab238fabb79c828006b2776353442f04ebb9966ae201cb04cbd209d45745d6dc2fff16593d19dc4efcbd42d6726fcad69ca95d480a47a55609cc37bb81ff14eddb8aaaf39f0b9a7c5cabdfcfb4f8763916671bf4c1c48fb731a41c04abd5c3cd3aa80f1ae6b620567b97de36a7165ed8a7a57c061197d406c55e24866cc58ed8d69afea0f5868975780ea47d75c4c9c95ce7d4367d5a3b05c8d548182020c02c21819355324cb0e7b0cf5c78be758b38a2d25d3ccbdacb11059d18e26355e0e16f42ccab1348e706f673c7572f916f928f7b13719de9ccceb4f7487faf4c20104215a162e3d1171f4b7e80065ee806c66804836f4d9693f812efb0e577d04339e5f2a9957dbd05299f3533370851d0f016e1c8feacac69498881bc12793101457ab4ee1bc2df54e74db09d66be91fca8d9c04d0b4d6cabf82ceed7b68af619db06d2fe3e35ff31b9dc52da326324ed306b70bded2c35f9d0ecc66ed378069298d600a5d3dec9be6ae19a2518b5679e8ed0a70418351b8198db716e616f70f4ad3ae8ed82adfff920411afa8823fe72f8d58294b7783b637d23fb209409b8e00b92ce39f15f4f8b8a6a7fc9f9ec15b4e091d1391632d914893c27e62b51d2294db9f6bb8d66b67b51ca664b959e7df0f8bd0b5315b78028799204235a4b01a314274a1c9f173e1c390dc97f25c6d8837c9085dbca1fa2ecafba9fdedf12ddc5e5eb3c441e6226d24d63ef8dd849505307074b04235c2ee76ea6a18849c7a811f0b1d1df72106adad2c524e79a66e57cbd78132a9e358e3c51b0b09441b495c3b82ed7d4aa40f9de7073f2cb5340a0b1d040e92a85e9c8c4a90fa4dd466be6cb072ad97a3748bda73026f3e72db3fdb7ceb48c30e412e4fc7fc56bc69cf5b46c0303f44cdace4deb15aa7e7565645218c32c71342899c429793cd8b53d2c3ac7481bedf4a60aa9e818867873e603f30c5dcf0ab422d260fa456fe776125c2911ce93e34acfbe6c049054dde92a23d5732d84919fbdbe9f3e1e05256b93eeb1a863ea6dc102faff6fd7765c2ec55d7f13d519fc12ae185254f3e734d01fea9e388cff25db5c7240b1a2d7d14e6761381d1829d0fb5ccc1db1a179cf8d4bdc82bdd6dda3855211098e701313f421effff3446d478ae8429dd0117b9d3f5f4f024a95591d5f1f75a3fcfc893cc6206f0cbc9c59899ac27c34609398eb897995b44ab7ba9456d7329730ca34a1fa5f78d1a8c037d2bd72f9b8b4c392f4df9076165bd03a34082242f74feb1c7b8d09d4403bdedcb5a9ca67c4be1b8e110ff691f557e3f8ed098db08391a1a8ccce3fba8bd8b9da1cfd7baa7c3fe1898032e2b741fa463f25c7bac3cea13ed447988a36979d0f07710cf0fabe2b6529d3534aa05ea4130b6ebe4df724356a353ba9ad5a72395cfe2d63cfb659852d90f694f39d634c13e7c71af2cfd2f13560d0432135a2a5847cb00129b458c91320225c17e548a5032afa8bc0314f1c536e941bf9fc387d55008dc2a7dbd21642fd16a964712d353a801be4734054a1f2a649ae93479a1edd9ddc132aa216ff7fa233ece8adcc5c5039244c77e4781dcb97e4afe7021b61004b27e91c1b3a09a6d7ec4abd4501132ac1236eb6caac2833d3a005dadcee7462168d15b4bc56cffa8704d04a60cec3aff74982d364c571d94b5d87c0c6813645b8c00b8155d249a622664a7a130aa45b2bbf871db765308fa0bff925392ac9a14939a2c2c6fa2d707c4c1f35d259fcc73c4d8f13dc1c7dd40a8415b8e540667dc19f0a8875f7a592b5f76894a97b72cb5d0d3f8ae0a7e027f2eb8bbd16ee29db78e25931cf17e85a5313d059c38e70dae5ebe681ccb0a1ee0797607fda23069a0032eb56b78302da95539f9d7c0cd4dcb3d5f2769d40d216f6bf6aa1df010d265e424cfcd1eca1fd63832ca3405e10009b90bfcf2d6e234b4e01b5e881599d8d158756e7dbb27b6eea771482d95f3fcfd2cfc8fdc209b45e28f62faacc8ebded2f14bae0e7d072ad54a3b6ad7dd032be04db8e652b917cca77eb6828dd8c0648f9a3668656f9fd1af4ab680e3a4818b697e05216209c8789da423a8e510cdfdb8e2bcc0eeb3d3115ad6b881353601ce6aecc45b715c42aa38d2a1cfa0a0b3d437516d3e2bd9c376dd5e4d0e50be719bf8b9fd12884857ed0ab1820ee9aef55b981f1e73e09351aaaefd67e86f864ad6002dacfd5a73f441f27ffd180f8dfd8baa7328f4fd0abde6fbe79c54db07c7e910253d410f42d9a1fa1bfcf39be38be4d3b8e6c83be2cd80f1f053a061a2c43c97479c8fdcf28a9c8ee6417e86e04c3afe375a2ca0ad5439041697936972c45223e46dbae25a5f65c1d93a2f4503a0a455171496324a5b6b27041e01a46bb7b5cdbc42e26a62b6203fb2c312fe05fdd4d0a3d8319c3ebed56b5e55019a00d8f937a6c83bf44a64114e8747ae2d5717d649daa909f5812fc2e491a234354da419563b1ef268e2b36b318b4dd51429617ea02ed5920a13dd399f8b2098cdb4090f2b5703f8beb4446303d068bfdaffffffffae2935f1dfd8a24aed7c70df7de3a668eb7a49b1319880dde2bbd9031ae5d82ffd601c11c2ca382d5acce6e1ac21723465f7f0cce4a55b67ae534777dbcd509dcfc635bee6ab49cecd45b5232479dfd097d9a840567bd82f158c4799532ecedee521b542ce86668464710649aea36d456380a887330ac517abefbeacb35ea77f5e3d1ee4cf86c89445e8ad2539858945c4d4252322fae17978714fd250bcc3b51972bb0171f142d5f2114e8cb6625979e3daeeb54754916cf5fb479359a7d924bca3248cfa3780a3c35bcec297fd96845d3809e591ad31741da9a4ca03530b2e8911b8c8b9a6570a8be8c99c7f45685646f62514a5eabf894285318a0c88164e435c2e965b72cd8e0bd9a80f101e29ebfbb96e8b2c0a8fd984d051fae2a7641ffebf244d10b16c39392d2b2d6cd6bee287ae7cba71321be3af43968faf0f89eb359815076925f5077c60decc05c6119c83e871747d5ec5011a5117c6dd0822082630b92dbf58028f8a49de434343051bd85da7c33626cfbc316920cc8f907470b66326b7b06a413df77b43b9e4b52cd172591fc5f3cacd7672b836b56a9e9bb984912d605d1c67d69459cb5e0f794c8fc73218f29cca0b09adb88b892f85957cac131e057b1ce841cca03e2d8bc20363e89c64e0b8ccd7ef3180011b4ac325db3fdf37778195de57e25b85726bdd1b5d4b4d7623be7d9259d19905700ecb55958c510102b64c7fb1b472517064067a054dce77c542498a65731e3b9e96491fff3014a7e44ce7a1aa170703992c2ee2fa496995e09f1fe9800f889e4b36bec144dd6716476c793abbf3506f489590b220899bb55daf9f108088b37006bb16bbea175589161d0212eabb1faac1761976841e197776cea2f6f8711ca1bfd6f534e3d3f0892d4d2dc45a636802a6863353fb33e27271087f5dd83bf9497ad546de846eb338e1d6701cad5d49e1ca018a23553c8c4efd95bf2cebd5f7c3eb419beff17f038c7d6eb085e5707ae4d141fe2f7db27be799abdfc308c5f551925f6814bf27d0be53a52dff338eb0dfc82cbd6a5b434ca25569925718e49c011a29f8bb329068ac66c334cf77b00298a3fb7eb0ad66c0e74392384a001311fe494d958d5a47792a171a989d294ba0635f077a836eddf953d53cc823c50c8f2ab76cfcedf88c178ede1800834bffe7840babadf70d07aa646955f26c70fe36bc200d0e9e75100c82c6be453cabf3341f9f782ea33b1cfbc74e0fe449d21f9e2aefff6b8a0fa9f68a0c29552799a43f9116fa0743b5509c48dbedbde770a0b3dc1a718b50ee7aa3a07cdf53676d72f60f68d63f4a98c0bd6019a0617caea969b5cd8c64b06a83158a0312752bf5580040c455cf2d23216dbd4d28b541dc225a7fb7238693ae936e8ddf965dd1de92cf824dbf8d43a488ee599d743887b68a96768cc87b1bf251b68f16bade6e73479078e3733feb05b83fc0e58ba3ae59881d75b6d34cbc2f26b12ab8c180dd1de31bad92774ce00bf2dd65fca9cbf8ae2e46897a6537d55fe5c80db0e9a4f576fe57538c9cd3b5f1b4bbcf4ab81a4ecdcab1e2f7b920c6ac98e895f8d23f38a9d6bd2af7cf6baf04483a36135735e77a7c65c3f56aacc55810602302a8abb1914ed2edebbba2a59f5e6a56e24a152008634dbaf9d09667beb2bf8c065cd523a2d50bbcd3509c9c7b34f743dbf5fbae7840ba6ec93ed62f937dd28bda4e137875acb0f27b2a4b3002df719f3096f3109a5126dab8177d948190fb969d1c7728f9e31ee5dc2efac364eff7eeb24080083b7b017c0a8331d59ba5500d7e8fb7f9c3a1479acaba081a449dd994ee6a5a6f660f74e51d5eb96acfa592bc446ef4aeecac9d9477fd30960cd7d2017038b64dba07e4c9761a05e824e8e0a646d94170611586cdab4472e8828abbdedde3098e8f8090a854974bd9dc090cb68e731b18b01bd364323161b7b6f30215bb49895562807f802ebb21fd2963a80e463605a56c75015579ded820dccfde02449407ff5bc8d9ed4e370ee51534f4bc64f1968b442182591f94060cefd5711ca86d0bd65fef5088d18f2ef48f8b438930dc5474f364e9586be5ecf25d987b113f6e30ffcfa56b4843bb62af51a6b8b281a6e962093f92a0c3418c6bb32305d32b2bd131dd334d9c67a833554838b23b0ca2dc24413ed07a13e70252a365644b90d713b69aca6ba6e7041ef411e142a947bbac7256f5aedd4f2d066f9b94946cd9b3564c0dd6564ca7d63f3c6749f2041c1aff6366435f229d4de16a919da01417bb11e03d0d07c056cde5fc4cc2ecb764c03ea639e776628a14904eb1dbb8fcf7967e639b2e9cee51f2a6303f8ddda794a1b8519702b480ac0acf584760b4caa780833ec7a4dcb4f7594cee509f9e258755202136c83be849265b93943a7fce2235bb7336d099a8dae4e82e51dac78fb8a9e791b8948fcb8b535cc1a3d75b7eb4b7244c168b060237cb7a20d7756603b8de61876e9cacfe55d90473501ade2f6c90cf6c2d6932b97f0afb23021f853cbf9a5e1c18976271fc8cfb1771220dc1e81ab0d4e18af7698ce7c43fa415b7232009510afa40a56059af3a3b4f97da134fc5003d3df7f17b35c1c0882bc4cc274993fe5092a3aca116192a0bc21dcfde28fc44ca1f46d34b2f43e146b8f72c1871b6065f66bb8cbe26f1a7791515f2c135148b97709fa4a8e20e66fa11e61c1be48564c3929a245cf58e57b47b7c40d4b16d954ed2229d915f29843474af0d1cdaed5867a63e61e9504a977de059df5aef4d9c0a22a598615c266c9ea1c32d1f3b382598768cc6ae38e92c1ef49df8c74b76740c367c89a6def18af15254462812f421b420851d98539ddc2c9cd0d318441628ee4636553fb1f46ab740cdef1fcade0ff443a67ae953de1ea21a4dba461fc558167af924f3a4454cc0aaa5b80f191d83e1c6e4c43d521162f0266ee22b1ec3f5d5d623b83e4e4f2ad388f735087b3fce3199e54af841be9b081f5b1c2ee77ad61dc7f6bf8746c875933f4d1227135513065be9b07a7d0d6418924cf2dc77ebac143bc8dd3b993f0e7f6ce5d30c0775fbb57b65c6cb57119024e42fcfb622c9af9afc6d4b6ee0ecd9fb3b8d65244b6ab52dabe14fd5d858e366557e17895228325f9ac9bd97a9ec4e405fe8963a19c3cd271d6b6a250929603f6b505cc37d5a576b2c8a761432f4eac0c3d22e0a8488f0c648c496de6086f6183f2aecc7212358badc0aabe0b21c9dce78b38a293e5d5e79a1261252b6b67a67b31c5f361b58b7df0bec0439ea41b3219a25e62705acfe6df86645a8a5e0761cf3c940e7655fb0c9efc0d86bc44a811a538d1016a8541d326eb66876b3bb7447af5de37c953edb728f51bc898157d2834878af270f292f3af65ae843365f3e4ee834fbcebf82e3469ac28b528d3f8b046746db0d74d91472705bdb6b57d201b4798cc0899b675c68df1dfba7e585fa67a73cf335686eec5e7419968e31747bebd66685c3d98a8d07edac34f89a8e9f3eeb3c383b4011d51c84919846e3859dc3c9aa100b987e3b3743d93307dd9cc98310ad8216378e31ca9ed5572b59db0a83f238e5cf2c8f1fe6dc1cf34f7b6e33c5e3f6d9388671bc171cccb4a96ad48cfe4cda74a98184e2a47dad7956fcbbb73fa879e836a90d69104a4cd599a6aa5e6281603a5e8e2aa535819fecb906ecf6af7e4482373f3b2f72098403c792e909d1a6ce4b6527bbd60917da1cf4eab5246548e7fc01186d98aba75e528416a4d00662205ce75dea781cc57f9883d885d9818d2b10172e5f935b3e3af71dbb7d46763498a5ae5d40c5920c040c3eb42a21697d031610de9de5ca2511ba1bdec68759761ff3c368cffabcbccc72c00cc23470857f39012ed2a115106dc29b721fc4227ae87c19df9b321b2d9187138e5e67c74992fc168ce83f08eb5210f12782919fe5d0bfb897cbdf5b31b113e508aab0c8077cf219f0f396344695ed9236b27105fbff7533c8ac48e879cf8a86b62dffbdef0e0f0e89890b452d5dbed35ef1135aeb062d9d169ebd9d37332238a7ab1e197e8c2605045041c0a8b801b11e407d8d682a5a93e23912c0d072ccc62725fb0cee3389243d11126c90409cd8ab22aeaf6d6a8a956456752554a2c0dcb1841f6586be363e6a9961096154349d6b646c626b0f30366d107ab547b6a57e79c98ff5692689245eb1be7ffb3256a7bf7ac915cdbbd74c760d75939992ba5cfc7d72e00cb8f02aef277e2e87d05b2648d5d0e4fd919fc500389f9caa2516211827a7c27f97761608ada7b6ea7694c5523b83f176e02933f089938826f0ad78946446df5f1186310ba8aff6b68e49047377c36de8fd53588c668d8fff6badb3f626e6e42ebeac03e165cfdeb9bca044b11238fa731d365252713eec3a3818160c7c1d98f837c13e00a3a5c521492e34b7a013acf2cd4341ed570a12156b0aa9e6584f36add45a306c49c83de0f288807164450ca83691efd9b74f53c5b71b570453dc1091a6f13f168f8e3cd59e488f613971b54af2146e68aa058f46ee9f7d6aeba4d0daab7b10fcabcbd73399e2cb58c1f796e3cb3ab52fdf02abdef9d744c937dc2883f6252cea37464c5244c70af888daf8eddd5d2188c4414c242ab14611788d7375d20a3d2fe5543634b92224a071b16a24a3f9a3a9543da30617f217127027f08945202b931507dd6095828ce19a4ab8b807ac1c18d7271aedcae6947c4897069be6e92c57eb5af4811f1da8c924cba12b05919ededd4c4a16c400cc7979f6b724b936263e09283d37da248b37d0f8899f9ef871d49d4bf7b6ebe1d969da45d1be8fb60d2391ba157b01fc5d17d201a383d8b8a70e35596ab05683e0c14af9c4ac1189281502bfecc9a1318f76b76dbe902a496a602f678adee9600c668fd94b878acc3995511aee40112849c02a293b45ad058f743e5c70f585cddf707dc6623962f33a40fddd71a6b0aca003293163cc08f06e97b206f61ffb9c4f09b0bb5e9870960a414b0a0c0e40f47b816f8f5ed6e1fd6dd6bd7172d13c646657e7e2573c5d7e0ff457e6dbca186574e70e64046100575ab89f30bd65a3d11e643d7fb1818eb80fba3fa4e911a4003dde5f16a0a628f5591307b8c897281e35317f9cbae543b809af48ca2bda385e066b7a5014d4074c1a35374f3ebf5733a54d744ac6cddda705c05d3a3fae5d09d74d1c1ff06bbccf4a966bedad8ff96746dfefac68c0b5023694ddddf0feb55d9a42d922891040c49a88196506c4a5fde5a1e311f211fc122b4d462e8c8e3923dd5111e9e545f065aec44ec938f24bf860ae947e27460d1600d262e365485c38858cf81a8aac15b22689bb47b6b5a7f14802ecbf5899b5d9220143ca38f459d778aeebe21216c0b8cb1d76912c671e10db49333942e2da2a291eb70b1cfd6e0425f73131a8eb705dff0af6c6334b5a92d01f6ef1f5625f3303fb7c4a6d61b8116f12e762116857b1695b80294030d5c14e4995328027afc6105cbe82df606fc61ec82486caa36afc7a72c014b70271f9bb386e4158c88f383965ead5b5ad770f58f31fad13f8574a08faa177aabacd4048036afe5b76ee7e16af84c3530da1b20745f33bbd50534cf72cee54078d3a953cbac52bcf11198d12eb86488a15c54045b1cc71444ffe6b633b0e600721b0489e12338a293e25850819a883e7dd996c18ceb97a198b71aa632c8960951794bdc253513554a523fe2385b7c5de676f9be752c4c78671b5e45f0c1e55027bccbcc64787d494e10f3636beb5a142bdce10db9465361b50076b927c0fe45cdc1bd9347575ef2b38ec232773f5b8f21c1899fd6e163f95d95174eede62acba89238ff89cb5d42f1abc07140b6c972d58e561cae9a6d2de06435194be35d4b021728bd30952c5c51cc6860814cc5075a590ae4d11235a2e38cba31da7a4dbb6c9d06b24481038e0f3fec61b2ec79c6167513e8e01f8a171536032d12273907e899c94afc86ae14f6f36c83a352ac025bd03fdd55af3e7b18c70c9dee38211a5359035b157a70eee2386ed2865698728263679af8ae1cfb5e95a242c687cbddf0d7f6048026709aa887b4b28113c43044baa7f31d1e86409c65cdbeb22d7967a8f2d0b497e19c14bd4e6f504f62b658348a3bdeb169a7d26c0d9e2e7468bba5630f162fee6175479d4ccce0cf90939aef49b6f3d6ecb8092f6bd02db0004b75ae5b4b90255abfe75ca283933247443ac053f092574306244221648991581b6f662b1a48a83819a4c6f00370df0b914b07c901a0039501dc91f66659a35b23f4b540f47859004404c102a060b9454914e52b687a798ae0619c61b21d3d934baa4efef1624df5cbe1b2fe3f91b21ab5a5cb9011e59b3b918289e26e71ed3537a7972b46d16dff407bd643cf90cb2545bb092617cd113b8efd6d9aa21e5d6da49b557b490b9edc069fb52c3fa02425a7d7670a37c4f5e0971ab973a500ec40c1c0dddd0a3064505c1121001a82a20fe69386a0a3000695afde7cb74e4513d8ca09bdab733838296e275de21961178dd06ac502ba0dd9fbe3169a917c6989ddc516ee40378646b3acfa3be65c710fc26a34f1ddf943894f2ef03aeac3772a1a8a163a5574b21d8170f7df6d1cf3e11f9e7a5c519be161ac76989fb680be811e648df554246f44aeaad99b4586f398d3c29a6077e510b62e0fc1187030da7bb0d0fa7362813b0727bebd898e0df398e47bc900af7721f5a40201770f4d91b323928117d7b93a261ddc7d122da211b76b0139b758311c796aac0cf890c35e97c083b718df5e862d7dca13a64abd510f2457805880439d1e54cfd6954a06183231a097ffe7f454aad4007c7c61930051bd04de754d2b7ebf21ae3a4296fd5cdd9c8c11b93349701845ca9723c3e1b1a61a349f9127f43010e0110a60a777bbcd5e1a0caf0390ff9df0ccef6188bdf158217dda7cd2e15a01ad10ca5683addee9f38d3be973d869b52422d7e06cc7f3060baad95d274512cc87599399b756047d0fcf81e178ae98b939b86931700811629586c9d80b91de1ee30cb4dd523b947e41e4fec4f31d6d536129795778fe72da7ea2ce14b9f6147aca51799febcbcdf1b646c929723ce678072ffdfe1b3a03d98ef5843fc828fd6b234d61303d3c90f58490b6f76d3f07b58b5247720c9e52bf5642148762ff5caf0b9755e6ad01436006f0b5227ff6a7b910c530eb610930ddd085254d6041954812d17a1e272dea17f53d2628a066453f7ddf4deb15b702eafdc9e5dfd2073a933efce183fce2006eb242fa03730855c7d6fcd333c70130af9944a05fbda2e423a90775619d0313232d0e7b94a9aa24977a4aae4c069101b7679c49cb9233bcb65514aa00fc4044ab1c8720894839279014c8d7b91bc4d39288c75bd69ebc8886653346d12069aec97b62a9faa100000b09164da9baa1025383638a72dfc3d1a2bfce102af98f964106a54fb052f11be3028382bb2cb1b13b29fa6cf5fc871f6d8ee540f79d252819360d04e92f9537efab1e0c581802c205527a5ac3d2b1c0fddadfea02dbb28362d5a6f0adbbee7f388088297fadfa7218a9fa72f58b9e39d1764c09eb15f33e6f4df16393ba9d9a858e498857072281a1a5e29216e1c8342bdb631c92acd6bf7f8911118beb2c28b7a3ec82593fe64163a31ea8604f2eb33b0b2047eacb593568381cd787b426d1bd13206d81f86081d8e966ae6d5d94632fb760489ccf36d1888c07807c12360a30ea6e569794dd21c5a5cbf62c9835f487db103be9224dd85535a9d2d49063892fee2b18d51e7af30b34f9f74e1893e7c36c1e580ae22b4cc8ca3fd80263383a63a2972faeab3de19ea3d67a0917e57cf417600eb754c152b0a28f3cfeb4fff9d115538a895bbe61a6b60b6d61f84bc98cbb1a9ae0d1ad176c21dfa4ccfd7f64442cb918b127e70026f6600c0183c778153d456caefad8b13cf5b3d76f6d45a895c211d58eca9c6225c713efed30eb773b1bf424061e7d82245203aa6e6567a879b9376fd30afd4322b25189b3b9182eba1436b9c395c27fddf2cb244d6de1823535ab2c66c25430b6cfa98c55ddc713261f8d797cc3e2d5664834e1b7647628d870c2315211fed278788a31efef59f9f5647897b883bd5779a074c38fb1d6deba45f077f5e18a10b03582bbed93148ac33e3d75cfa9f14dc767d124c91b9c83dd1c1aa3f3d558f1692bf83f3c6f84f4d777810bbf2f674c1a5c3bba55eebba1b4b837e89d1af4e27f7713f683187f1638e2f7663b8cac0d8ae39d65580e0d91610c6d081fe67cb1b078563c27622018776e3b49b226f68979e367f35096aa4910fd8b8085954b239f9f96ef9a5c29b633b30e2d7ff1777dac3edcd43f26e9b95ed852f814f571124eee9547def0746be562495b8cfa1fc7524529fd9bf602e82fd475d775a32560730b89acb610efde6fcc272c98266dda804ba20a72bf8d6ae6c15f0e93b77ea0380470591f6356d39100c04d89cd081546e2a962e540d3aac136dbc986a34d62d2aca6306335da72bf8d53d0165c5b020b2bfdb945df34678eefa0918b9de0d2e25eeb67b5599174b583b529fe70249c21e95b9c7cea63606511138f29d78e0350e3f7ed3a3f2971b5282c021f61e5117bc9c5e632ce4e375736231109a79be24c12468f78e85294121738e3ff3915d94dfbf4ab8d5d41011e28898610a43a534efe7d7792ca1600d9a654d997c4b27ecd1e93fcafbb3f2f5c5833e3a3446c431f2de104c070a9a5b567bca3e92c00c843b5a452faa1b102be1d19e22b52afc362dcbd83cf5a7c42cff56545cfefd76d283d20e2a17e0258a2293835d1d618f2fa0d1da224ef54964c89e0b270cfd622a53c0454858bb0e2aba47039ea365c33353802c8ca1226e0dc7bd4f6c96e5cb266ef4a4582175bf706930c658de7d5d24c69d8a9689a9f3d55327a96cfa4219140f168cd9c82d48c729a2b48f60e204025c9e91953697ea61542b7debd56374ff78b0f73636d70ca09d273942f62ea926e0b24bb1ede163c8a6b4456229b13fb12ca484fa9e85b99124f967223bb43b3edf81d3795d68fd9b1f419779637e6e7724c2611e8d230c07345ea6db5cb851257bd48c3b9c33e5028ffbc2e8a43b685dad9acb2541d1f273aef451f873540b3ccfc64ed9ffa53fa1d27a0ede1d544cb28917a66b527727ef917997cc52453e0f6b86273ecb5d912f3a3ca7da63deefcbf349406d496d1aa666f5c27d0ce40ea47423b16ab9edd37b9f7d4df62f0e77ec94a00682b6fb460360f7049d7ade9c2e05944865f3125f2281e1f2a69e5f9ebf1afbb3b16123ae0a602fe04420af1eab854c15a9fd75592e3ea7870fb86be4ffe217d335a7a7d277591504cab6f4761c1eef64fba967a8b3bc2741bcf110809237044a82ed4f7f047b068b8eea6aa42632f66e0fe832ad41995c3b5ccba0bdfa477f1c680591e765f084f4d89dc069ba0be647f11f7e9669c04868bdba62e54927a96876c79c734bdd6d142441f4c608a179ca765fd4270f17bc0d2a5484d5d3697da7af4624175b1726316277554d09d7803da5ccd4c33926587ea92199ccc76f14efa74e21bbf1154a6e81c3fce2b5a7f2ca3671785d8896aa023958ac6c913c1081298002d9cdb1a6f7ef01dfffa25fa603807d478423673b33c145c09948baa1e1b61d42d652f6392b5b16e9e9bb402b13751b05f9cfbfc6436032d87d56380977470c324d29f91da6a90a84a42bd299b807662f1baea38635d31ca991fb2f620ae6c2c7f84331f6a1a16696ebbc99c0dacb43df6851833b9cb6c02e325c01e1e029157d8697079c222ed50ba0892c2372b0971d3d2d275ac95d9801fbe3da7136cf9696fe88122561a91b32620efb177e3269bdcfa4fb28536dfda8132508459f5ac0071731c4a7c0e141172684f531c9115080cc7410c74b46df45d372d2249e2124b33edbfa034427adb88f0a97f7bbbc8848bc44ff60531c76205304354167b6c2075549cb414b62c10434a126eab9683ec89df4969fe70f8ee90e8a732c5a48823f54394735d38648e5179cb8a4a7e33038ed2976d3bb6fd6f8c2223ee986e9c2d4813120e3417144a2dfbcf8b2be853992fe6ada74cf12c698e784f0493316369b010fef5f8fb613d8fa11db6fe1639b3cb2f7cc5f1e69015301607aa13f35a79949c061a9090ab43c2d02b76d3ead65438ee86a60eb2367c7d792c95848735c57fd279a6a27a6b95b21102059a75facfcf8a78c8118bfb3bb715e799b8dba8a746fb2f3a65f9b29e7b220eed3fd3ed5ad04eccf294cc1da017a12b998498f47b92d7ad5905e1a02848097ec2c95a35d9bd3aaaec038d82b70ab4f57a5d7e4ddc94c789c912be156fc422881f174b4033699cde2c844e6152460d7f40a4526c03bd9b5e32a260ed1d12e06995692d5a193aecde0061affc73d9f2148c0fc51d3786b86cfd5337bb8c8d1298668a390340d131293ead0e2078efb0b4bf865963a794cbec09780ad52ff14334f5fd4ff70373338379c6f14737443abd27b90d90a0b588e217152e2b15c050a")]).await?;
-
-    sleep(std::time::Duration::new(2, 0)).await;
-
-    let sapling_activation_tree = TreeState {
-        network: String::from("main"),
-        height: 663150,
-        hash: String::from("0000000002fd3be4c24c437bd22620901617125ec2a3a6c902ec9a6c06f734fc"),
-        time: 1576821833,
-        sapling_tree: String::from("01ec6278a1bed9e1b080fd60ef50eb17411645e3746ff129283712bc4757ecc833001001b4e1d4a26ac4a2810b57a14f4ffb69395f55dde5674ecd2462af96f9126e054701a36afb68534f640938bdffd80dfcb3f4d5e232488abbf67d049b33a761e7ed6901a16e35205fb7fe626a9b13fc43e1d2b98a9c241f99f93d5e93a735454073025401f5b9bcbf3d0e3c83f95ee79299e8aeadf30af07717bda15ffb7a3d00243b58570001fa6d4c2390e205f81d86b85ace0b48f3ce0afb78eeef3e14c70bcfd7c5f0191c0000011bc9521263584de20822f9483e7edb5af54150c4823c775b2efc6a1eded9625501a6030f8d4b588681eddb66cad63f09c5c7519db49500fc56ebd481ce5e903c22000163f4eec5a2fe00a5f45e71e1542ff01e937d2210c99f03addcce5314a5278b2d0163ab01f46a3bb6ea46f5a19d5bdd59eb3f81e19cfa6d10ab0fd5566c7a16992601fa6980c053d84f809b6abcf35690f03a11f87b28e3240828e32e3f57af41e54e01319312241b0031e3a255b0d708750b4cb3f3fe79e3503fe488cc8db1dd00753801754bb593ea42d231a7ddf367640f09bbf59dc00f2c1d2003cc340e0c016b5b13"),
-        orchard_tree: String::from(""),
-    };
-
-    connector.add_tree_state(sapling_activation_tree).await?;
-
-    let first_transaction_state = TreeState {
-        network: String::from("main"),
-        height: 663173,
-        hash: String::from("0000000001b80f69c0b4a02228b59f2793ba9a4c3e8742fe56b9c9e331a6005d"),
-        time: 1576823459,
-        sapling_tree: String::from("018310f940d4b52399e656d8a090bd87f64a4abe80fffb2b4a882675def41f943c0114dd70a8a8a22869785b86b1dcad5e9508f419aad84ef8e83d50ec061117022310000199517be06af7c07c2d393c590c62add4dbcd9cc3402167521786f91a5d01d538000001989561014441f9f9043e11c01e220730df2219c090fa02f58d278fb7f447271601fa6d4c2390e205f81d86b85ace0b48f3ce0afb78eeef3e14c70bcfd7c5f0191c0000011bc9521263584de20822f9483e7edb5af54150c4823c775b2efc6a1eded9625501a6030f8d4b588681eddb66cad63f09c5c7519db49500fc56ebd481ce5e903c22000163f4eec5a2fe00a5f45e71e1542ff01e937d2210c99f03addcce5314a5278b2d0163ab01f46a3bb6ea46f5a19d5bdd59eb3f81e19cfa6d10ab0fd5566c7a16992601fa6980c053d84f809b6abcf35690f03a11f87b28e3240828e32e3f57af41e54e01319312241b0031e3a255b0d708750b4cb3f3fe79e3503fe488cc8db1dd00753801754bb593ea42d231a7ddf367640f09bbf59dc00f2c1d2003cc340e0c016b5b13"),
-        orchard_tree: String::from(""),
-    };
-
-    connector.add_tree_state(first_transaction_state).await?;
-
-    let second_transaction_state = TreeState {
-        network: String::from("main"),
-        height: 663187,
-        hash: String::from("00000000027fa4bfd6c012325a44eef7211a6162b5979507f07603333e9b3068"),
-        time: 1576824317,
-        sapling_tree: String::from("018d5b2b12a89cabbeb6c98bde09fbea143d49ea8e4dbf1d612e4906e73e1af96b001000018b01c7e7b2b183d022fc35e351e6423aee8885debc899036d0bc3b389c9f161501dba0595ce728b41452a9c595341074cf01e1152abe401db2b30a9ab007ad006e0001989561014441f9f9043e11c01e220730df2219c090fa02f58d278fb7f447271601fa6d4c2390e205f81d86b85ace0b48f3ce0afb78eeef3e14c70bcfd7c5f0191c0000011bc9521263584de20822f9483e7edb5af54150c4823c775b2efc6a1eded9625501a6030f8d4b588681eddb66cad63f09c5c7519db49500fc56ebd481ce5e903c22000163f4eec5a2fe00a5f45e71e1542ff01e937d2210c99f03addcce5314a5278b2d0163ab01f46a3bb6ea46f5a19d5bdd59eb3f81e19cfa6d10ab0fd5566c7a16992601fa6980c053d84f809b6abcf35690f03a11f87b28e3240828e32e3f57af41e54e01319312241b0031e3a255b0d708750b4cb3f3fe79e3503fe488cc8db1dd00753801754bb593ea42d231a7ddf367640f09bbf59dc00f2c1d2003cc340e0c016b5b13"),
-        orchard_tree: String::from(""),
-    };
-
-    connector.add_tree_state(second_transaction_state).await?;
-
-    let third_transaction_state = TreeState {
-        network: String::from("main"),
-        height: 663189,
-        hash: String::from("00000000025a03b6b11364b9c01a5fc9a927eca98ae899bb88f0ddf2de77fda3"),
-        time: 1576824596,
-        sapling_tree: String::from("01f1c5f08a96b8e0befe04ff061dd51ed96ffcc7fcd96d011dc06255692989731c001001ea6dbe4e95ec503600ee3260105ffcc2ceb63eb34ec323c794eebfc49b7beb2c018b01c7e7b2b183d022fc35e351e6423aee8885debc899036d0bc3b389c9f161501dba0595ce728b41452a9c595341074cf01e1152abe401db2b30a9ab007ad006e0001989561014441f9f9043e11c01e220730df2219c090fa02f58d278fb7f447271601fa6d4c2390e205f81d86b85ace0b48f3ce0afb78eeef3e14c70bcfd7c5f0191c0000011bc9521263584de20822f9483e7edb5af54150c4823c775b2efc6a1eded9625501a6030f8d4b588681eddb66cad63f09c5c7519db49500fc56ebd481ce5e903c22000163f4eec5a2fe00a5f45e71e1542ff01e937d2210c99f03addcce5314a5278b2d0163ab01f46a3bb6ea46f5a19d5bdd59eb3f81e19cfa6d10ab0fd5566c7a16992601fa6980c053d84f809b6abcf35690f03a11f87b28e3240828e32e3f57af41e54e01319312241b0031e3a255b0d708750b4cb3f3fe79e3503fe488cc8db1dd00753801754bb593ea42d231a7ddf367640f09bbf59dc00f2c1d2003cc340e0c016b5b13"),
-        orchard_tree: String::from(""),
-    };
-
-    connector.add_tree_state(third_transaction_state).await?;
-
-    let fourth_transaction_state = TreeState {
-        network: String::from("main"),
-        height: 663194,
-        hash: String::from("000000000096e4f59faae478c0ea5cb982c1cba223a92d15fbcf74795a0b4e5d"),
-        time: 1576824759,
-        sapling_tree: String::from("01f1c5f08a96b8e0befe04ff061dd51ed96ffcc7fcd96d011dc06255692989731c001001ea6dbe4e95ec503600ee3260105ffcc2ceb63eb34ec323c794eebfc49b7beb2c018b01c7e7b2b183d022fc35e351e6423aee8885debc899036d0bc3b389c9f161501dba0595ce728b41452a9c595341074cf01e1152abe401db2b30a9ab007ad006e0001989561014441f9f9043e11c01e220730df2219c090fa02f58d278fb7f447271601fa6d4c2390e205f81d86b85ace0b48f3ce0afb78eeef3e14c70bcfd7c5f0191c0000011bc9521263584de20822f9483e7edb5af54150c4823c775b2efc6a1eded9625501a6030f8d4b588681eddb66cad63f09c5c7519db49500fc56ebd481ce5e903c22000163f4eec5a2fe00a5f45e71e1542ff01e937d2210c99f03addcce5314a5278b2d0163ab01f46a3bb6ea46f5a19d5bdd59eb3f81e19cfa6d10ab0fd5566c7a16992601fa6980c053d84f809b6abcf35690f03a11f87b28e3240828e32e3f57af41e54e01319312241b0031e3a255b0d708750b4cb3f3fe79e3503fe488cc8db1dd00753801754bb593ea42d231a7ddf367640f09bbf59dc00f2c1d2003cc340e0c016b5b13"),
-        orchard_tree: String::from(""),
-    };
-
-    connector.add_tree_state(fourth_transaction_state).await?;
-
-    sleep(std::time::Duration::new(2, 0)).await;
-    connector.apply_staged(663_200).await?;
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_simple_sync() {
-    let mut rng = rand::thread_rng();
-
-    let num: u32 = rng.gen_range(0..100000);
-    let temp_dir = TempDir::new(&format!("dwld_test_{num}")).unwrap();
-    let path = temp_dir.path().to_path_buf();
-
-    let darkside_server_uri =
-        zingoconfig::construct_lightwalletd_uri(Some(format!("http://127.0.0.1:9067")));
-
-    prepare_darksidewalletd(darkside_server_uri).await.unwrap();
-
-    let server_id = zingoconfig::construct_lightwalletd_uri(Some(format!("http://127.0.0.1:9067")));
-
-    let light_client = ClientManager::new(
-            server_id,
-            path,
-            "still champion voice habit trend flight survey between bitter process artefact blind carbon truly provide dizzy crush flush breeze blouse charge solid fish spread"
-        )
-            .build_new_faucet(663150, true).await;
-
-    let result = light_client.do_sync(true).await.unwrap();
-
-    println!("{}", result);
-    assert!(result.has_key("result"));
-    let JsonValue::Object(res_value) = result
-        else { panic!("Expected object, got {result:?}") };
-
-    assert_eq!(res_value["result"], "success");
-    assert_eq!(res_value["latest_block"], 663200);
-    assert_eq!(res_value["total_blocks_synced"], 50);
-}
-
-#[tokio::test]
-async fn reorg_away_send() {
-    let mut rng = rand::thread_rng();
-
-    let num: u32 = rng.gen_range(0..100000);
-    let temp_dir = TempDir::new(&format!("dwld_test_{num}")).unwrap();
-    let path = temp_dir.path().to_path_buf();
-
-    let darkside_server_uri =
-        zingoconfig::construct_lightwalletd_uri(Some(format!("http://127.0.0.1:9067")));
-
-    prepare_darksidewalletd(darkside_server_uri).await.unwrap();
-
-    let server_id = zingoconfig::construct_lightwalletd_uri(Some(format!("http://127.0.0.1:9067")));
-
-    let light_client = ClientManager::new(
-            server_id,
-            path,
-            "still champion voice habit trend flight survey between bitter process artefact blind carbon truly provide dizzy crush flush breeze blouse charge solid fish spread"
-        )
-            .build_new_faucet(663150, true).await;
-
-    light_client.do_sync(true).await.unwrap();
-    println!("{}", light_client.do_balance().await.pretty(4));
-}
